ft_printf bientot fini mais sans les flags

+++ srcs/ft_intlen.c
@@ -6,7 +6,7 @@
 /*   By: zminhas <zminhas@student.s19.be>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/01 17:03:36 by zminhas           #+#    #+#             */
-/*   Updated: 2020/12/02 17:26:44 by zminhas          ###   ########.fr       */
+/*   Updated: 2020/12/06 16:49:07 by zminhas          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int		ft_intlen(int n)
 {
-	int		len;
+	unsigned long int	len;
 
 	len = (n < 0) ? 1 : 0;
 	while (n /= 10)
